requirements.txt
panel
pyautogen
openai
supabase
